requirements.txt
chromadb
langchain
BeautifulSoup4
gpt4all
langchainhub
pypdf
chainlit
python-dotenv

transformers
torch
torchvision
torchaudio
datasets
evaluate
rouge_score
peft
loralib
numpy
